move "sutup" scripts to internal
