+++ README.md
@@ -6,6 +6,4 @@
 
 Директория `cmd` содержит файлы - "точки входа" для приложения.
 
-Директория `pkg` содержит "общие" библиотеки (которые могут быть использованы в сторонних приложениях).
-
 Директория `internal` содержит файлы с бизнес-логикой приложения.
