+++ internal/setup/setup.go
@@ -1,0 +1,12 @@
+package setup
+
+import (
+	"log/slog"
+	"os"
+)
+
+func Init() {
+	LoadEnv()
+	SetLogLevel(os.Getenv("LOG_LEVEL"))
+	slog.Info("setup was completed")
+}
